requirements.txt
fastapi
uvicorn[standard]
pyyaml
httpx
psycopg2-binary
